requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.1
django-phonenumber-field==3.0.1
entrypoints==0.3
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.17
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
phonenumberslite==8.10.14
pluggy==0.12.0
psycopg2==2.7.6.1
py==1.8.0
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
pytest==4.6.2
pytest-base-url==1.4.1
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-selenium==1.16.0
pytest-variables==1.7.1
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
zipp==0.5.1
